Complete code for sum multiples of 5 or 3 function AboutApplying...js
Comment recommended chaining _.range and _.reduce, which I did, but without
wrapping them in _.chain and _.value.  Tried it both ways, both were
successful.  Chose the less verbose format, then stumbled upon a method for
this using filter between range and reduce.  I did not use filter in my
solution.  Perhaps mine is not the best but it's the one I created.